+++ src/main.cpp
@@ -245,3 +245,4 @@ extern "C"
 	__declspec( dllexport ) DWORD AmdPowerXpressRequestHighPerformance = 1;
 }
 #endif // _WIN32
+